# pkg/trace/agent: reduce concurrency in Process method