This change removes all goroutines from the `Process` method of the trace agent to improve speed, simplify logic.

Benchmark results:
```
name                old time/op    new time/op    delta
Process/defaults-4    60.4µs ± 0%    49.9µs ± 0%  -17.40%

name                old alloc/op   new alloc/op   delta
Process/defaults-4    23.2kB ± 0%    23.2kB ± 0%   -0.00%

name                old allocs/op  new allocs/op  delta
Process/defaults-4       265 ± 0%       265 ± 0%    0.00%
```

The first commit in the PR sets up the benchmark and the second one removes the test-only code and the goroutines, attaching the result above.