@gbbr 2019-02-06
I'm marking this WIP and putting it on hold, since I'm leaving on PTO. An alternative idea is to send the entire list of traces received in a payload to the processor, causing each payload to be treated in its own routine. This will remove the scenario where traces are dropped entirely.

I will explore this option when I get back.