@erichgess 2019-02-06
Take with a grain of salt since I don't know this code nearly as well as you.  And this may already be what you're thinking.  But I was thinking about this last night and from what I've understood:

1. The original code was inefficient because it was breaking off goroutines for CPU bound tasks which don't benefit from concurrency, so we were just adding the cost of goroutine management and getting no parallelization.
2. The a.tracePkgChan <- &tracePkg ties to an IO bound goroutine consumer (if I understand correctly), which makes this line blocking.

So could we just move the write to `a.tracePkgChan` to a goroutine, so we break out the smallest block of logic which is needed for the IO bound functionality and benefitting from concurrency there while keeping the CPU bound code together and not adding goroutine overhead?