@gbbr 2019-02-06
Your idea is good, and it's one option, which is close to our current solution in `master`, and better. But we might be able to do even better...

Another idea would be to make a bigger scale change in the HTTP receiver which currently [iterates through each trace](https://github.com/DataDog/datadog-agent/blob/b4d1e4cc2759dd771a745643d0205ec961afe846/pkg/trace/api/api.go#L233) in a payload, [sending them down a channel](https://github.com/DataDog/datadog-agent/blob/b4d1e4cc2759dd771a745643d0205ec961afe846/pkg/trace/api/api.go#L253) to the processing function and [dropping them](https://github.com/DataDog/datadog-agent/blob/b4d1e4cc2759dd771a745643d0205ec961afe846/pkg/trace/api/api.go#L257) if the processing function is slow. Instead, I'd like to experiment with dropping this loop and it's channel, dumbing down the HTTP receiver to have only one purpose: pre-sample, receive and decode payloads. After it decodes them, it could send them to be processed within the same goroutine (unique to each server request in Go) by means of a callback. This way, each payload (and contained trace) will be processed in its own goroutine. This way, traces will never be dropped and each goroutine can take its time to process its payload. Then, the pre-sampler can reject payloads when memory/CPU usage is too high, just like it does now. This should serve as a great improvement in terms of how much traffic the trace agent can take, as well as make memory _a lot_ more manageable. It will also make things easier to test and isolate responsibilities better.

I hope my explanation makes a bit of sense.

This feels like it could be a great solution and it's something I'd like to experiment with, possibly as a continuation of this PR.