@erichgess 2019-02-06
I think it does, I believe that's what I did in the X-Ray converter.  Each
request is run within it's own goroutine from decoding, to processing, to
sending to Trace Edge and within that Request Handler goroutine there are
no sub goroutines, just the sequence of steps to process and send and deal
with retries on the send etc.  Let me know if that's similar to what you're
thinking.  If so, I can attest that the design works really well, the code
is much easier to understand and maintain and performs very well for us.

On Wed, Feb 6, 2019 at 9:26 AM Gabriel Aszalos <notifications@github.com>
wrote:

> Your idea is good, and it's one option, which is close to our current
> solution in master, and better. But we might be able to do even better...
>
> Another idea would be to make a bigger scale change in the HTTP receiver
> which currently iterates through each trace
> <https://github.com/DataDog/datadog-agent/blob/b4d1e4cc2759dd771a745643d0205ec961afe846/pkg/trace/api/api.go#L233>
> in a payload, sending them down a channel
> <https://github.com/DataDog/datadog-agent/blob/b4d1e4cc2759dd771a745643d0205ec961afe846/pkg/trace/api/api.go#L253>
> to the processing function and dropping them
> <https://github.com/DataDog/datadog-agent/blob/b4d1e4cc2759dd771a745643d0205ec961afe846/pkg/trace/api/api.go#L257>
> if the processing function is slow. Instead, I'd like to experiment with
> dropping this loop and it's channel, dumbing down the HTTP receiver to have
> only one purpose: pre-sample, receive and decode payloads. After it decodes
> them, it could send them to be processed within the same goroutine (unique
> to each server request in Go) by means of a callback. This way, each
> payload (and contained trace) will be processed in its own goroutine. This
> way, traces will never be dropped and each goroutine can take its time to
> process its payload. Then, the pre-sampler can reject payloads when
> memory/CPU usage is too high, just like it does now. This should serve as a
> great improvement in terms of how much traffic the trace agent can take, as
> well as make memory *a lot* more manageable. It will also make things
> easier to test and isolate responsibilities better.
>
> I hope my explanation makes a bit of sense.
>
> This feels like it could be a great solution and it's something I'd like
> to experiment with, possibly as a continuation of this PR.
>
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub
> <https://github.com/DataDog/datadog-agent/pull/2995#issuecomment-461041487>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ADboG-6_EGqid0uq6xkvfUXdEIx0YRJyks5vKuYwgaJpZM4ajXgB>
> .
>


-- 
- Erich Ess