@gbbr 2019-02-06
You're correct, that's what I'm thinking. I'm glad to hear this method has a proven record. I'm looking forward to trying it out.