@erichgess 2019-02-06
I actually simplified it even further.  The golang `http.ListenAndServe`
already handles concurrent requests, so I just let it manage the actual
goroutine per request.  This helped simplify the code an extraordinary
amount.

On Wed, Feb 6, 2019 at 10:10 AM Gabriel Aszalos <notifications@github.com>
wrote:

> You're correct, that's what I'm thinking. I'm glad to hear this method has
> a proven record. I'm looking forward to trying it out.
>
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub
> <https://github.com/DataDog/datadog-agent/pull/2995#issuecomment-461057689>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ADboG2sibkxtWByaenyUrGJNndQ8ypOTks5vKvBYgaJpZM4ajXgB>
> .
>


-- 
- Erich Ess