@gbbr 2019-03-01
This is a great PR to reference as proof that reducing concurrency in the Process function indeed improves performance. Closing and readdressing processing concurrency in follow-up work.